requirements.txt
djangorestframework
channels
channels_redis
